feat: support fp32 grad accumulations

+++ src/llama_recipes/policies/mixed_precision.py
@@ -1,6 +1,3 @@
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# This software may be used and distributed according to the terms of the Llama 2 Community License Agreement.
-
 import torch
 
 from torch.distributed.fsdp.api import (
@@ -11,7 +8,7 @@ from torch.distributed.fsdp.api import (
 fpSixteen = MixedPrecision(
     param_dtype=torch.float16,
     # Gradient communication precision.
-    reduce_dtype=torch.float16,
+    reduce_dtype=torch.float32,  # Use float32 for gradient communication. (Llama-3, Megatron_LM like)
     # Buffer precision.
     buffer_dtype=torch.float16,
 )
@@ -19,7 +16,7 @@ fpSixteen = MixedPrecision(
 bfSixteen = MixedPrecision(
     param_dtype=torch.bfloat16,
     # Gradient communication precision.
-    reduce_dtype=torch.bfloat16,
+    reduce_dtype=torch.float32,  # Use float32 for gradient communication. (Llama-3, Megatron_LM like)
     # Buffer precision.
     buffer_dtype=torch.bfloat16,
     cast_forward_inputs=True,
@@ -27,7 +24,7 @@ bfSixteen = MixedPrecision(
 
 bfSixteen_mixed = MixedPrecision(
     param_dtype=torch.float32,
-    reduce_dtype=torch.bfloat16,
+    reduce_dtype=torch.float32,  # Use float32 for gradient communication. (Llama-3, Megatron_LM like)
     buffer_dtype=torch.bfloat16,
 )
